Moved run_simulations to separate file
